requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
pycparser==2.23
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
